+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Dungeon 3D".
 - Tests for "The grand festival - II".
 - Tests for "Quaternion multiplication".
+- Tests for "Maze /w teleporters and jumps".
 
 ## [1.12.0] - 2022-09-01
 ### Added